Block size per export The block size can be configured per export. In the exportd config file the bsize parameter tells the block size exports = (    {eid = 1; bsize = 0;root = "/home/jean-pierre/rozoClients/mdev/tests/export_1"; md5=""; squota=""; hquota=""; vid=1;},    {eid = 2; bsize = 1;root = "/home/jean-pierre/rozoClients/mdev/tests/export_2"; md5=""; squota=""; hquota=""; vid=1;},    {eid = 3; bsize = 2;root = "/home/jean-pierre/rozoClients/mdev/tests/export_3"; md5=""; squota=""; hquota=""; vid=1;},    {eid = 4; bsize = 3; root = "/home/jean-pierre/rozoClients/mdev/tests/export_4"; md5=""; squota=""; hquota=""; vid=1;} );
0 is  4K
1 is  8K
2 is 16K
3 is 32K

+++ rozofs/rpc/epproto.h
@@ -32,7 +32,7 @@ typedef struct epp_status_ret_t epp_status_ret_t;
 struct epp_estat_t {
 	uint32_t eid;
 	uint32_t vid;
-	uint16_t bsize;
+	uint32_t bsize;
 	uint64_t blocks;
 	uint64_t bfree;
 	uint64_t files;
